Some changes

+++ app/model/listing.py
@@ -4,7 +4,7 @@ from typing import Optional
 class Property(BaseModel):
     property_id: Optional[int] = None
     title: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     summary: Optional[str] = None
     price: Optional[int] = None
     location: Optional[str] = None
